fixes tests and update packages
