+++ package.json
@@ -1,16 +1,3 @@
 {
-  "private": true,
-  "devDependencies": {
-    "babel": "^6.0.15",
-    "babel-cli": "^6.0.15",
-    "babel-jest": "^15.x.x",
-    "babel-preset-es2015": "^6.0.15",
-    "babel-preset-react": "^6.0.15",
-    "jest-cli": "^15.x.x",
-    "react-addons-test-utils": "^0.14.1",
-    "react-dom": "^0.14.2",
-    "lerna": "2.0.0-beta.28",
-    "eslint": "^3.x.x",
-    "eslint-plugin-react": "^6.x.x"
-  }
+  "private": true
 }
