+++ packages/react-layout-builder/.babelrc
