+++ packages/react-layout-builder/package.json
@@ -1,53 +1,66 @@
 {
-    "name": "react-layout-builder",
-    "version": "1.0.0",
-    "description": "",
-    "main": "build/index.js",
-    "author": "code@alvinsj.net",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/alvinsj/react-layout-form"
-    },
-    "keywords": [
-        "reactjs",
-        "form",
-        "layout",
-        "builder",
-        "react-component"
+  "name": "react-layout-builder",
+  "version": "1.0.0",
+  "description": "",
+  "main": "build/index.js",
+  "author": "code@alvinsj.net",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/alvinsj/react-layout-form"
+  },
+  "keywords": [
+    "reactjs",
+    "form",
+    "layout",
+    "builder",
+    "react-component"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/alvinsj/react-layout-form/issues"
+  },
+  "homepage": "https://github.com/alvinsj/react-layout-form",
+  "peerDependencies": {
+    "react": "^15.x.x"
+  },
+  "dependencies": {
+    "object-assign": "^4.x.x",
+    "form-serialize": "alvinsj/form-serialize#export-fn"
+  },
+  "devDependencies": {
+    "babel": "^6.5.x",
+    "babel-cli": "^6.16.x",
+    "babel-jest": "^15.x.x",
+    "babel-preset-es2015": "^6.16.x",
+    "babel-preset-react": "^6.16.x",
+    "enzyme": "^2.4.1",
+    "eslint": "^3.7.x",
+    "eslint-plugin-react": "^6.x.x",
+    "jest": "^15.1.1",
+    "lerna": "^1.1.3",
+    "react-addons-test-utils": "^15.3.2",
+    "react-dom": "^15.3.2"
+  },
+  "scripts": {
+    "jest": "NODE_PATH=src:NODE_PATH jest",
+    "test": "jest",
+    "test:watch": "npm run jest -- --watch",
+    "compile": "babel -d build/ src/",
+    "prepublish": "npm run compile"
+  },
+  "jest": {
+    "automock": true,
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
+    "testPathDirs": [
+      "<rootDir>/__tests__"
     ],
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/alvinsj/react-layout-form/issues"
-    },
-    "homepage": "https://github.com/alvinsj/react-layout-form",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "peerDependencies": {
-        "react": "^15.x.x"
-    },
-    "dependencies": {
-        "object-assign": ""
-    },
-    "scripts": {
-        "jest": "NODE_PATH=src:node_modules jest",
-        "test": "npm run jest",
-        "compile": "babel -d build/ src/",
-        "prepublish": "npm run compile"
-    },
-    "jest": {
-        "automock": true,
-        "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-        "testPathDirs": [
-            "<rootDir>/__tests__"
-        ],
-        "testPathIgnorePatterns": [
-            "pending"
-        ],
-        "unmockedModulePathPatterns": [
-            "react",
-            "node_modules"
-        ],
-        "modulePathIgnorePatterns": []
-    }
+    "testPathIgnorePatterns": [
+      "pending"
+    ],
+    "unmockedModulePathPatterns": [
+      "react",
+      "node_modules"
+    ],
+    "modulePathIgnorePatterns": []
+  }
 }
