+++ packages/react-layout-form/.babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["es2015", "react"]
+}
+
