+++ packages/react-layout-form/package.json
@@ -1,55 +1,64 @@
 {
-  "name": "react-layout-form",
-  "version": "1.0.0",
-  "description": "A form layout generator for your React components",
-  "main": "build/index.js",
-  "author": "code@alvinsj.net",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/alvinsj/react-layout-form"
-  },
-  "keywords": [
-    "reactjs",
-    "form",
-    "layout",
-    "builder",
-    "react-component"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/alvinsj/react-layout-form/issues"
-  },
-  "homepage": "https://github.com/alvinsj/react-layout-form",
-  "dependencies": {
-    "classnames": "^2.1.3",
-    "object.assign": "^4.0.1",
-    "react-layout-builder": "^1.x.x"
-  },
-  "devDependencies": {
-    "enzyme": "^2.4.1"
-  },
-  "peerDependencies": {
-    "react": "^15.0.0"
-  },
-  "scripts": {
-    "jest": "NODE_PATH=src:node_modules jest",
-    "test": "npm run jest",
-    "compile": "babel -d build/ src/",
-    "prepublish": "npm run compile"
-  },
-  "jest": {
-    "automock": true,
-    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-    "testPathDirs": [
-      "<rootDir>/__tests__"
+    "name": "react-layout-form",
+    "version": "1.0.0",
+    "description": "A form layout generator for your React components",
+    "main": "build/index.js",
+    "author": "code@alvinsj.net",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/alvinsj/react-layout-form"
+    },
+    "keywords": [
+        "reactjs",
+        "form",
+        "layout",
+        "builder",
+        "react-component"
     ],
-    "testPathIgnorePatterns": [
-      "pending"
-    ],
-    "unmockedModulePathPatterns": [
-      "react",
-      "node_modules"
-    ],
-    "modulePathIgnorePatterns": []
-  }
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/alvinsj/react-layout-form/issues"
+    },
+    "homepage": "https://github.com/alvinsj/react-layout-form",
+    "dependencies": {
+        "classnames": "^2.1.3",
+        "object.assign": "^4.0.1",
+        "react-layout-builder": "^1.x.x"
+    },
+    "devDependencies": {
+        "enzyme": "^2.4.1",
+        "babel": "^6.5.x",
+        "babel-cli": "^6.16.x",
+        "babel-jest": "^15.x.x",
+        "babel-preset-es2015": "^6.16.x",
+        "babel-preset-react": "^6.16.x",
+        "eslint": "^3.7.x",
+        "eslint-plugin-react": "^6.x.x",
+        "jest": "^15.1.1",
+        "lerna": "^1.1.3"
+    },
+    "peerDependencies": {
+        "react": "^15.0.0"
+    },
+    "scripts": {
+        "jest": "NODE_PATH=src:NODE_PATH jest",
+        "test": "jest",
+        "compile": "babel -d build/ src/",
+        "prepublish": "npm run compile"
+    },
+    "jest": {
+        "automock": true,
+        "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
+        "testPathDirs": [
+            "<rootDir>/__tests__"
+        ],
+        "testPathIgnorePatterns": [
+            "pending"
+        ],
+        "unmockedModulePathPatterns": [
+            "react",
+            "node_modules"
+        ],
+        "modulePathIgnorePatterns": []
+    }
 }
